migration fixing

+++ database/migrations/2022_06_29_122126_create_business_partners_table.php
@@ -17,7 +17,7 @@ class CreateBusinessPartnersTable extends Migration
             $table->id();
             $table->unsignedInteger('business_id');
             $table->unsignedInteger('taxpayer_id');
-            $table->double('shares')->nullable();
+            $table->decimal('shares')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
